pyproject.toml
[project]
name = "fnirsi-ps-control"
version = "0.1.0"
description = "PC control library and CLI for the FNIRSI DPS-150 regulated power supply"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.11"
keywords = ["fnirsi", "dps-150", "power-supply", "serial", "usb-cdc"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
  "pyserial>=3.5",
  "typer>=0.12",
  "rich>=13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-pyserial",
]

[project.scripts]
fnirsi = "fnirsi_ps_control.cli:app"

[project.urls]
Homepage = "https://gitlab.com/yourname/fnirsi_ps_control"
Repository = "https://gitlab.com/yourname/fnirsi_ps_control"
"Issue Tracker" = "https://gitlab.com/yourname/fnirsi_ps_control/-/issues"

# ---------------------------------------------------------------------------
# Build
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fnirsi_ps_control"]

# ---------------------------------------------------------------------------
# Ruff – linting & formatting
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=fnirsi_ps_control --cov-report=term-missing"
